Refactor constants out of src/lib.rs into its own file
