+++ src/client.rs
@@ -1,7 +1,7 @@
 use crate::{
+    constants::{MAINNET_GENESIS_HASH, SOCKET_NAME},
     ledger::account::Account,
     state::summary::{SummaryShort, SummaryVerbose},
-    MAINNET_GENESIS_HASH, SOCKET_NAME,
 };
 use clap::{Args, Parser};
 use futures::{
