+++ tests/canonicity/blocks.rs
@@ -2,10 +2,10 @@ use crate::helpers::setup_new_db_dir;
 use mina_indexer::{
     block::{parser::BlockParser, BlockHash},
     canonicity::{store::CanonicityStore, Canonicity},
+    constants::{MAINNET_CANONICAL_THRESHOLD, MAINNET_GENESIS_HASH, PRUNE_INTERVAL_DEFAULT},
     ledger::genesis::GenesisRoot,
     state::IndexerState,
     store::IndexerStore,
-    MAINNET_CANONICAL_THRESHOLD, MAINNET_GENESIS_HASH, PRUNE_INTERVAL_DEFAULT,
 };
 use std::{fs::remove_dir_all, path::PathBuf, sync::Arc};
