+++ tests/command/store.rs
@@ -2,11 +2,13 @@ use crate::helpers::setup_new_db_dir;
 use mina_indexer::{
     block::parser::BlockParser,
     command::{signed::SignedCommand, store::CommandStore},
+    constants::{
+        CANONICAL_UPDATE_THRESHOLD, MAINNET_CANONICAL_THRESHOLD, MAINNET_GENESIS_HASH,
+        MAINNET_TRANSITION_FRONTIER_K, PRUNE_INTERVAL_DEFAULT,
+    },
     ledger::genesis::parse_file,
     state::IndexerState,
     store::IndexerStore,
-    CANONICAL_UPDATE_THRESHOLD, MAINNET_CANONICAL_THRESHOLD, MAINNET_GENESIS_HASH,
-    MAINNET_TRANSITION_FRONTIER_K, PRUNE_INTERVAL_DEFAULT,
 };
 use std::{fs::remove_dir_all, path::PathBuf, sync::Arc};
